Move shopify-deploy.js -> @shopify/slate-sync
