+++ packages/slate-sync/prompt-if-published-theme.js
@@ -2,9 +2,8 @@
 const chalk = require('chalk');
 const inquirer = require('inquirer');
 const https = require('https');
-const YAML = require('yamljs');
 const slateEnv = require('@shopify/slate-env');
-const {event} = require('@shopify/slate-analytics');
+const { event } = require('@shopify/slate-analytics');
 const config = require('../config');
 const figures = require('figures');
