+++ packages/slate-tools/package.json
@@ -22,6 +22,7 @@
     "@shopify/slate-env": "^1.0.0-alpha.24",
     "@shopify/slate-liquid-asset-loader": "^1.0.0-alpha.24",
     "@shopify/slate-tag-webpack-plugin": "^1.0.0-alpha.24",
+    "@shopify/slate-sync": "^1.0.0-alpha.24",
     "@shopify/themekit": "0.6.11",
     "archiver": "^2.1.0",
     "autoprefixer": "6.7.7",
